backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
openai>=1.54.0
cryptography==41.0.7
python-dotenv==1.0.0
sqlparse==0.4.4
openpyxl==3.1.2
httpx>=0.27.0

